job-platform-backend-main/requirements.txt
Django==4.2.16
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
whitenoise==6.7.0
gunicorn==22.0.0
